Solution to Input and Print in Hexadecimal and Octal

+++ Input and Print in Hexadecimal and Octal@@6788d5012db25efec946f941/code.c
@@ -2,8 +2,8 @@
 
 
 int main() {
-    float a;
-    scanf("%f",&a);
+    int a;
+    scanf("%d",&a);
     printf("Hexadecimal: %X\n", a);
     printf("Octal: %o\n", a);
     return 0;
